backend/requirements.txt
fastapi
uvicorn
paho-mqtt
pymysql
